+++ src/filter/error.filter.spec.ts
@@ -1,12 +1,12 @@
 import { ArgumentsHost, BadRequestException } from "@nestjs/common";
 
 import { ErrorFilter } from "./error.filter";
-import { InvalidCredentialsError } from "../exception/invalid-credentials.error";
-import { InvalidOtpError } from "../exception/invalid-otp.error";
-import { InvalidRefreshTokenError } from "../exception/invalid-refresh-token.error";
-import { InvalidResetPasswordRequestError } from "../exception/invalid-reset-password-request.error";
-import { UserExistsError } from "../exception/user-exists.error";
-import { UserNotFoundError } from "../exception/user-not-found.error";
+import { InvalidCredentialsError } from "../error/invalid-credentials.error";
+import { InvalidOtpError } from "../error/invalid-otp.error";
+import { InvalidRefreshTokenError } from "../error/invalid-refresh-token.error";
+import { InvalidResetPasswordRequestError } from "../error/invalid-reset-password-request.error";
+import { UserExistsError } from "../error/user-exists.error";
+import { UserNotFoundError } from "../error/user-not-found.error";
 
 describe("ErrorFilter", () => {
   let filter: ErrorFilter;
