+++ src/user/user.service.spec.ts
@@ -3,12 +3,12 @@ import { User, UserRole } from "../entities";
 import { IPostRegisterUserHandler } from "./interfaces/post-register-user-handler.interface";
 import { IPreRegisterUserHandler } from "./interfaces/pre-register-user-handler.interface";
 import { IUserValidator } from "./interfaces/user-validator.interface";
-import { InvalidCredentialsError } from "../exception/invalid-credentials.error";
-import { InvalidUserError } from "../exception/invalid-user.error";
+import { InvalidCredentialsError } from "../error/invalid-credentials.error";
+import { InvalidUserError } from "../error/invalid-user.error";
 import { MockFactory } from "mockingbird";
 import { Repository } from "typeorm";
 import { Test } from "@nestjs/testing";
-import { UserExistsError } from "../exception/user-exists.error";
+import { UserExistsError } from "../error/user-exists.error";
 import { UserFixture } from "../../test/fixtures/user/user.fixture";
 import { UserService } from "./user.service";
 import { faker } from "@faker-js/faker";
